Fix childless redirects
Makes parseNodes consider the Command of a redirect target when no
contents can be read. For example, given a "foo" literal with a non-null
Command, and a "bar" literal forwarding to the "foo" node, the parse
results' context contains the command of the "foo" node, and the parsed node
is the "bar" node.

This follows the same logic as a regular redirect.
`CommandDispatcherTest#testExecuteRedirectedMultipleTimes` highlights
the parse tree contains the redirected node (`redirectNode`), while
the executed Command is that of the target node (`concreteNode`).

Fixes https://github.com/Mojang/brigadier/issues/46