Check for permissions when producing suggestions
Currently, you can get suggestions for commands you do not have
permission to execute. For each candidate node, we first check if
canUse returns true for the source. If so, we also check the
context-aware canUse variant also returns true. In this case we need to
construct a temporal CommandContextBuilder with a "fake" range, since we
do not know the real range the considered node takes up in the original
input. Otherwise, we would need to re-parse the argument, which is
expensive.

The
getCompletionSuggestions_argument_impermissibleContext_with_no_argument
test case shows that in order to be totally sure a source can use a
node, we must perform two double permission checks. The first, done by
the #parse method, passes the parsed arguments. The second one is
performed by #getCompletionSuggestions and does not know about the
arguments. Still, the context objects are different and the
context-aware predicate may return different results for each case.